style: separate ESLint configurations
Instead of dumping every rule configuration into one big file, we'll
modularise them by plugin instead, and extend those in our main config
file. This enables better organisation and gives us more control over
where the rules are applied.
