+++ configs/eslint/deprecation.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+    plugins: [
+        'deprecation',
+    ],
+    rules: {
+        'deprecation/deprecation': 'error',
+    },
+};
