+++ configs/eslint/eslint-comments.cjs
@@ -1,0 +1,9 @@
+module.exports = {
+    extends: ['plugin:eslint-comments/recommended'],
+    rules: {
+        'eslint-comments/no-unused-disable': 'warn',
+        'eslint-comments/disable-enable-pair': ['warn', {
+            allowWholeFile: true,
+        }],
+    },
+};
