+++ configs/eslint/no-constructor-bind.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    plugins: ['no-constructor-bind'],
+    rules: {
+        'no-constructor-bind/no-constructor-bind': 'error',
+        'no-constructor-bind/no-constructor-state': 'error',
+    },
+};
