+++ configs/eslint/no-use-extend-native.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+    plugins: ['no-use-extend-native'],
+    rules: {
+        'no-use-extend-native/no-use-extend-native': 'error',
+    },
+};
