+++ configs/eslint/promise.cjs
@@ -1,0 +1,13 @@
+module.exports = {
+    extends: [
+        'plugin:promise/recommended',
+    ],
+    rules: {
+        'promise/catch-or-return': ['warn', {
+            allowFinally: true,
+        }],
+        'promise/always-return': ['error', {
+            ignoreLastCallback: true,
+        }],
+    },
+};
