+++ configs/eslint/sonarjs.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+    extends: ['plugin:sonarjs/recommended'],
+    rules: {
+        'sonarjs/no-duplicate-string': 'error',
+    },
+};
